+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Karafka framework changelog
 
+## Unreleased
+- Commit stored offsets upon rebalance revocation event to reduce number of messages that are re-processed.
+- Support cooperative-sticky rebalance strategy.
+- Replace offset commit after each batch with a per-rebalance commit.
+- User instrumentation to publish internal rebalance errors.
+
 ## 2.0.11 (2022-09-29)
 - Report early on errors related to network and on max poll interval being exceeded to indicate critical problems that will be retries but may mean some underlying problems in the system.
 - Fix support of Ruby 2.7.0 to 2.7.2 (#1045)